Use Luigi built-in configuration and fix ArrayExpress downloader
This is a progressive set of changes that will make a generally better usage of
Luigi for running various components of the pipeline.

The SRR download is now parallelized within Luigi tasks instead of GNU
parallel.

Add a generic DownloadExperiment task so that other step in the pipeline can
disregard the data source.

Add a luigi-wrapper script for end-users.

Add an example configuration.