Add support for 10x BAM submissions to SRA
Detect which pipeline branch to take by looking up the assay type of a
dataset. Add a special case for FAC-sorted single-cell datasets that
should be treated as bulk.

Add support for 10x BAM SRA submissions. This is done by looking up the
header of the BAM files to infer the sequencing layout and calling
bamtofastq downstream on the original submission.

Temporarily use the branch of bioluigi with improved sratools support
and Cell Ranger.